requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
boxsdk[jwt]
python-dotenv